체력 및 게임 win die

+++ Assets/02.Script/Multi/NetworkSpawnManager.cs
@@ -21,7 +21,7 @@ public class NetworkSpawnManager : MonoBehaviour
 
     PhotonView PV;
 
-    float RandomRange;
+    int RandomRange;
 
     private void Start()
     {
